Move user validation to separate file
Also renamed the corresponding test file.
