+++ internal/auth/user_test.go
